requirements.txt
python-telegram-bot>=21.6
youtube-transcript-api>=1.2.0
openai>=1.40.0
python-dotenv>=1.0.0
yt-dlp>=2025.1.15
